+++ pkg/api/mock/matcher/common_test.go
@@ -12,13 +12,18 @@ type testObject struct {
 }
 
 func (o *testObject) EndpointURL(ctx context.Context) (*url.URL, error) { return nil, nil }
+func (o *testObject) GetIdentifier(context.Context) (string, error)     { return o.Identifier, nil }
 
 type testObject2 struct {
 	Identifier string `anxcloud:"identifier"`
 }
 
 func (o *testObject2) EndpointURL(ctx context.Context) (*url.URL, error) { return nil, nil }
+func (o *testObject2) GetIdentifier(context.Context) (string, error)     { return o.Identifier, nil }
 
 type testObjectNotStruct string
 
 func (o *testObjectNotStruct) EndpointURL(ctx context.Context) (*url.URL, error) { return nil, nil }
+func (o testObjectNotStruct) GetIdentifier(ctx context.Context) (string, error) {
+	return string(o), nil
+}