+++ pkg/apis/core/v1/location_genclient.go
@@ -2,14 +2,15 @@ package v1
 
 import (
 	"context"
+	"fmt"
 	"net/url"
-	"strings"
 
 	"go.anx.io/go-anxcloud/pkg/api"
 	"go.anx.io/go-anxcloud/pkg/api/types"
 )
 
-// EndpointURL returns the default URL for core location operations
+// EndpointURL returns the default URL for core location operations.
+// It supports Get by-code if `Code` is set and `Identifier` is not.
 func (l *Location) EndpointURL(ctx context.Context) (*url.URL, error) {
 	op, err := types.OperationFromContext(ctx)
 	if err != nil {
@@ -21,20 +22,24 @@ func (l *Location) EndpointURL(ctx context.Context) (*url.URL, error) {
 		return nil, api.ErrOperationNotSupported
 	}
 
-	return url.Parse("/api/core/v1/location.json")
+	endpointSuffix := "location.json"
+	if op == types.OperationGet && l.Identifier == "" && l.Code != "" {
+		endpointSuffix = "location/by-code.json"
+	}
+
+	return url.Parse(fmt.Sprintf("/api/core/v1/%s", endpointSuffix))
 }
 
-// FilterRequestURL rewrites the request URL to use the /location/by-code.json endpoint
-// when Get operation by Code is requested
-func (l *Location) FilterRequestURL(ctx context.Context, url *url.URL) (*url.URL, error) {
+// GetIdentifier returns the objects identifier
+func (l *Location) GetIdentifier(ctx context.Context) (string, error) {
 	op, err := types.OperationFromContext(ctx)
-	if err != nil {
-		return nil, err
+	if l.Identifier != "" || err != nil {
+		return l.Identifier, nil
 	}
 
-	if op == types.OperationGet && l.Identifier == "" && l.Code != "" {
-		url.Path = strings.Replace(url.Path, "location.json", "location/by-code.json", 1)
+	if op == types.OperationGet {
+		return l.Code, nil
 	}
 
-	return url, nil
+	return "", nil
 }